Lending Risk & Automation Platform/requirements.txt
fastapi
uvicorn
httpx
tenacity
pydantic
pandas
polars
scikit-learn
xgboost
shap
imbalanced-learn
sqlalchemy
pytest
mlflow
joblib
pyyaml
matplotlib
pyarrow
